alex/llm_web_agent/requirements.txt
streamlit
openai
langchain
langgraph
python-dotenv
pandas
rapidfuzz
langchain-community
langchain-openai